+++ CHANGELOG.md
@@ -2,14 +2,17 @@
 
 ## Current version
 
+### 2.6.1 - Logger file-access guard
+- fix: LogWriter concurrent file write race
+
+## History
+
 ### 2.6.0 - Platform detection & version compatibility
 - add: Enums.PlatformOS
 - add: Assembly.PlatformInfo
 - extension for: AssemblyInfo.ToString/ToFullString
 - extension for: Version.RequiresUpdate
 - extension for: Test runner counter/colors
-
-## History
 
 ### 2.5.1 - Severity extension
 - extension for: Severity (Trace, Verbose, Any, All)